renamed file to firebaseInit from adminInit
