+++ server/src/actions/deleteConnectedUser.ts
@@ -1,5 +1,5 @@
 // Delete a ConnectedUser document given a document's index. This should typically be a socketId, but it can also be something else.
-import { connectedUsersCollection } from '../utilities/adminInit'
+import { connectedUsersCollection } from '../utilities/firebaseInit'
 
 export const deleteConnectedUserByUID = async (socketID: string) => {
   try {
