+++ server/src/utilities/firebaseInit.ts
